Update patch set 2
Patch Set 2:

> > I think something is relaying on py_sems
 > 
 > Oh, nice catch, thought it was also disabled in debian/rules but I
 > see it is not. Ok I'll verify the resulting .debs against the ones
 > we've got in NGCP to see if there's any other missing modules or
 > similar.

Ok I've been looking into this and this is another complete mess. :(

So py_sems relies on a python-sip module (which has nothing to do with normal SIP stuff, it's just a C/C++ bindings generator). The py_sems code has bit rotten substantially, and it has not built for some time, from jenkins it seems to fail already in mr4.5 and mr5.0 builds (https://jenkins.mgm.sipwise.com/view/NGCP%20projects/job/sems-binaries/6/consoleFull), so if something relied on that, well it's already broken. 

I've started looking into fixing the code, but I'd have to learn the whole SIP language thingy, which might take time, and I don't think fixing py_sems should be a blocker for this set of fixes TBH, which are an improvement over the current situation and certainly not a regression?

Patch-set: 2
Label: Code-Review=0