TT#26264 Add systemd unit service file
One reason to add it is to have consistency when bringing up and shutting down
the system, or when restarting services, to have more integration with the init
system as a whole.

But apart from that reason, it has been noticed as part of TT#31163 (service
action tests) that the autogenerated unit file in
/run/systemd/generator.late/ngcp-sems.service does not properly depend on
MariaDB/MySQL (has "After=mysql.service", but not Wants or Requires).

Unlike with sysvinit, with systemd the services are started in parallel, so
starting "after" another doesn't mean that the former is finished and fully
working by the time that the latter is ready.

Under these tests, this service sometimes fails to start, so it might be because
the DB is not ready (or at least it can be one of the reasons contributing to
the failure).  In any case, if it should start after "mysql.service" in sysvinit,
SEMS probably assumes that "mysql.service" would be fully working, so the
additional annotation "Requires" is more correct than having only
"After=mysql.service".

The commands in "ExecStartPre" could probably be moved elsewhere, but this
change is trying to preserve the behaviour of /etc/init.d/ngcp-sems until (or
unless) we decide to move it elsewhere.

Change-Id: I46b0b7b92fc6b13b467aafbb38e9383e88d0ad65