TT#26264 Add systemd unit service file, remove sysvinit support
One reason to have full systemd support is to increase consistency when bringing
up and shutting down the system, or when restarting services, to have more
integration with the init system as a whole.

But apart from that reason, the original reason is that it has been noticed as
part of TT#31163 (service action tests) that the autogenerated unit file in
/run/systemd/generator.late/ngcp-sems.service does not properly depend on
MariaDB/MySQL (it had "After=mysql.service", but not Wants or Requires).

Unlike with sysvinit, with systemd the services are started in parallel, so
starting "after" another doesn't mean that the former is finished and fully
working by the time that the latter is ready.

Under these tests, this service sometimes fails to start, so it might be because
the DB is not ready (or at least it can be one of the reasons contributing to
the failure).  In any case, if it should start after "mysql" in sysvinit (or
"mariadb.service", as we're renaming it under systemd), SEMS probably assumes
that "mariadb.service" would be fully working, so the additional annotation
"Requires" is more correct than having only "After=mariadb.service".

As part of this change and since we are revamping the init support, we also
remove sysvinit support (including /etc/default/ngcp-sems) and delete or modify
some maintainer scripts referencing old sysvinit support.

Change-Id: I46b0b7b92fc6b13b467aafbb38e9383e88d0ad65