Update patch set 1
Patch Set 1:

> Black Magic... Voodoo people...

I've read in the man page that this will work differently after compat level 10 (same as the others that Guillem mentioned related with installing init).  So we'll have to change it a bit in 1-2 years.

But it's a neat solution for the time being :)

Will merge now.

Patch-set: 1
Label: Code-Review=0