TT#34414 Changing systemd unit file TimeoutStopSec to 10
The default is 90s, so when it fails for some reason, it takes (unnecessarily)
too long to recover.

This happens when calling twice "ngcp-service ngcp-sems restart", or "systemctl
restart ngcp-sems", while having debugging enabled.  The processes apparently
get blocked, until they get killed after 90s of default timeout.

Perhaps it is still necessary (and better) to fix it in other ways, like finding
the root cause of the blockage (racing conditions during initialisation?).
Waiting for ~1s between restarts almost completely solves the problem, which
points to some problem of that sort.

However, given the current problem and the fact that 10s should be enough to for
this service to stop, it's probably a good idea to change from the default to
10s, so if finding similar problems for other reasons, they have less impact and
the timeouts are more acceptable.

Change-Id: I5487493bee5c9147972d9fa6e063176a8d9ca4fd