TT#36053 Make systemd sems service require redis-master
This service requires write access to redis, and relies on redis-master
to setup Redis in that mode. If we switch the redis master/slave roles
while SEMS is still running, or while we are trying to shut it down,
it gets very confused and takes very long to realize and shutdown.
So we should always make sure that redis-master is running before we
start sems, and stop sems before we stop redis-master.

Change-Id: Iaa43186230df6ab8365c94585192fbb6db2652cc