TT#154255 Remove redis dependency from unit file
We are migrating from redis to keydb so for now we need to support both
so we should not depend on specific key-value storage in unit file but
should use database.key_value.flavor value in override file.

But we can't redefine dependencies in override file just add additional
ones. So remove it from unit file.

Change-Id: I94ef109ce68c61ba3d195d00527e2d0612b6d886