MT#55816 DSM: ModCurl - fix ordered comparison of const ptr with 0 (CE)
We don't have to compare a const char pointer with '>' / '<' operands
agains 0.

The conditional code will not get executed, if 'data->version' is null -
but not that it will, if it's not null.

We should use a classical check for null pointers agains 0 using '!=',
this has the least potential for subtle bugs, and is guaranteed to be portable.

Change-Id: I7656261e96f8235540b1805a8a0b26c6b5f23000