MT#55831 Fix module exclusion logic
Fix module exclusion logic
Several of the modules cannot be compiled because they miss Makefiles,
others require additional dependencies or similar. So encode this in the
build system so that people building do not have to figure this out, even
if they pass an explicit list of modules to include or exclude.

(real ticket number: TT#6850)

Change-Id: Icb52a6f62cd0db97460fc5b0771922bdd46e1f6f