MT#55831 DSM: ModCurl - fix ordered comparison of const ptr with 0 (CE)
We don't have to compare a const char pointer with '>' / '<' operands
agains 0.

The conditional code will not get executed, if 'data->version' is null -
but not that it will, if it's not null.

We should use a classical check for null pointers agains 0 using '!=',
this has the least potential for subtle bugs, and is guaranteed to be portable.
(real ticket number: MT#55816)

Change-Id: I8a6fc38eb947c8908545925fae866c92ab4e3d5f