MT#56321 Treat 183 as 200OK (other leg), when early media forced
There is a need to add an exception processing for 183 Session Progress,
and the following in-dialog requests/responses, in case the 183
has been previously sent to the B2B with the hf
'P-Early-Announce: force'.

This adds the following behavior.

When 183 is received, and the caller has been updated with the
new media capabilities, according to those needed to embed early
media into the media session with the caller (via re-INVITE), then:
- even though the 183 is treated similarly to 200OK in terms of media
  updates, do not send ACK to the leg going towards DSM,
  becausethe sems-b2b giving the DSM, it's still in stage INVITE/183
- do not re-negotiate the leg going towards the DSM,
  after the caller has been updated with the new media (a usual behavior)
  because the sems-b2b giving the DSM still considers the dialog here
  in the Early stage
- do not set the leg going towards DSM into the Connected state,
  because by the fact, it's still in the Early stage, and setting of
  it into the Connected state, will break processing of BYE / CANCEL.
- upon receiving the BYE from the caller (after the DSM announce is heard),
  answer right away with 200OK, but do not forward BYE to the let going
  towards DSM. Instead initiate the CANCEL towards it, because this leg
  is technically still in the Early stage of the dialog.

Original ticket number: TT#187351

Change-Id: Id6e05202add1bcbd358eecbcd5e2cbda1a995b32