MT#56321 AmOfferAnswer: Re-use existing SDP for 183 with no body
Re-use existing local sdp, in case there is a sequential 183 coming,
but it has no SDP.

This will fix the cases, when:
- first 183 has been sent, and provided to the system the SDP
- second/third etc. 183 is coming, but has no SDP body

To not confuse the recipient (end subscriber) of these 183,
the media description in the scenario with 183 having no SDP body,
mustn't be changed (session id, media and attributes).

Previously the behavior was to generate local SDP with
a new session id and add sems-b2b into the media processing by that
(new media port, codecs and other attributes).

However, this change doesn't have an influence on very first 183
sent with no SDP body, because there was essentially no local SDP
saved before. So in this situation behavior is not changed.

Change-Id: I53dd57ae8a4739d5356d94fff45fb290896fd777