MT#56321 Add 'Allow:' header to the 200OK, if present in the request
We have to take care of keeping the actual 'Allow:' header in the
request, if the incoming request has had it.

One of the use cases for that is for example the INVITE coming
to the AA, when we must respond with the 200OK,
but there is no 'Allow:' header inside the 200OK.

Certain systems can be confused by that, even though the RFCs are
not forcing us to do so. Which can even lead to an impossibility to use
a normal call clearing with the BYE method, from behalf of the caller.

Change-Id: I9bafb89e4aa5e53077450068f847c07a497b43f1