MT#56321 DSM: Finishing the DSM playback, support: 486, 487 and 603
We have to add more SIP response codes, which will be checked against
the isDSMPlaybackFinished(). So that when the DSM playback is finished,
the call is properly released on the caller's side.

Exceptionally 487 is added, for cases like:
a call to HG, where nobody answers and the timeout triggers
a cancelation of all the legs.

Change-Id: I1f7cf81dd5966843713fc41c21ea0249174c13ab