MT#56321 Implement 'recvonly' beginning on-hold
Improvement of the way we detect, whether or not the
received SDP offer is the on-hold request.

Now we do not take into account the 'recvonly' cases,
when the offer arrives with that
(only 'inactive', 'sendonly' or "zeroed" on-hold).

From now on we also want to consider 'recvonly' as
the completely equal on-hold request, which expects
that the end recipient of that will return us the 'sendonly'
(or worst case 'sendrecv').

Original ticket number: 0055934

Change-Id: Ice3b767c4aef2c6e3e96199219ad0d0e04e93ecf