MT#56321 Force Early Announce: add support for the `Proceeding` state
We have to treat 18X replies coming from DSM (usually it's 183)
in `Proceeding` state on the same manner as we do for the `Early` state.

This is because it affects entirely the following call flow after that,
and has to do with a proper clearing of the dialog, when receiving
4XX class of messages (which stop playback and have to terminate this
session: 4XX -> BYE conversion).

Additionally: set the status to the `Disconnected` for the 4XX class
response, which ends up the DSM playback.

Original ticket's number: 56806

Change-Id: I3b0f3ff360cd774fcc7862d34e75747454660d9a