MT#57417 'WebRtcSpl_NormU32' is static but used in inline function 'log2_Q10_B'
This is a fix for the following warning:

     entropy_coding.c:101:11: warning: 'WebRtcSpl_NormU32' is static but used in inline function 'log2_Q10_B' which is not static
       101 |   zeros = WebRtcSpl_NormU32( x );
           |           ^~~~~~~~~~~~~~~~~
     entropy_coding.c:92:10: warning: 'kRPointsQ10' is static but used in inline function 'stepwise' which is not static
        92 |   return kRPointsQ10[ind];
           |          ^~~~~~~~~~~

This happens because the inline function has inside it a call
to the static function.

To fix the warning, we just make the `log2_Q10_B()` and `stepwise()` functions
static. It's quite safe, since they are not even in use by anything.

Change-Id: Id6588d85c7d347d5669d6495188e798fe7c70b05