MT#56184 build: Stop using INSTALL_TOUCH before installing the targets
This was a workaround for Solaris specific issues in the install program
there. The problem is that this can end up generating empty files, that
will not be detected by the packaging system as it will find the file is
present on destination and thus does not need to be installed.

Change-Id: Ib6107957cbf03c4e7da6cf2998d9103224f674d3