MT#57254 dsm/mod_redis: print ephemeral TCP port (connect/disconnect)
The hiredis library does not provide a direct way
to retrieve the allocated port number.

We hav to retrieve the socket file descriptor (redis_context->fd)
from the `redisContext` structure and using the `getsockname()`
get the local address associated with that socket.
Then the `sin_port` field of the resulting `sockaddr_in` structure
will contain the allocated ephemeral port.

By using `ntohs()`, just convert the port number from network
byte order to host byte order for printing. However make an additional
on-the-go type casting into the `unsigned int` also, to be able to print.
(unsigned int is guaranteed to be at least 16 bits,
so this is not a lossy conversion).

Change-Id: Id2a8e9ee4fad6f28db099323fb68fd2db0bfafc2