MT#40962 increase idle sleep time
If there are no timers then it makes no senes to process ticks in real
time. Sleep up to 0.5 seconds in that case. When timers are added we
are immediately woken up through the conditional variable. Only setting
the shutdown flag would not wake up the thread, but a sleep time of half
a second should be an acceptable delay for shutdowns.

Analogous to 695d902841

Change-Id: I37adf078825470af99e6a7755df3d8786d3eeaa6