MT#60408 Answer with 'inactive' when 'inactive' requested during the on hold
This commit affects cases when the on-hold is requested,
and the MoH emulation on the SEMS is additionally used.

Currently regardless the method used by the one who sets the call
on-hold, the answer will be always 'a=sendrecv', which leads to
a premature termination of the call at the RTPEngine
(because RTPEngine takes the 'a=sendrecv' into account and
terminates the call due to no seeing the audio)

We need to make sure, that an originator of the on-hold, is being
answered:
- with the 'a=recvonly', if an originator requests the 'a=sendonly'
  method of the on-hold to be used.
- with the 'a=inactive', if an originator requests the 'a=inactive'
  method of the on-hold to be used.

And after the on-hold is restored (by 'a=sendrecv'),
make sure to send the SDP answer with the 'a=sendrecv' as well.

Change-Id: I50df81be16a75d9e813484012938e2a94ec422e1