MT#60480 Limitation of origin id/ver to 64bits (unsigned long long)
SDP session origin id and version are limited by it's size
to 64 bits (8 bytes), that is because we are using `unsigned long long`
type for both id and version. Hence the maximum value it can hold
is two sixty-forth power, so equals: 18,446,744,073,709,551,615

In case one intentionally sets it to something like:
"o=- 18446744073709551615 18446744073709551615 IN IP4 192.168.0.1"

SEMS gets overwhelmed with that conversion inside the code make
it just to a string literal 'F'.

To overcome it, just use __uint128 (which is of 128bits size)
for the session origin id and version.

Change-Id: I2ad9659aa81dad79969749053dc3fd0d69e2cbd2