MT#62063 Recvonly cannot put on hold
Previously introduced by 175b00d72b79561d7
the `Recvonly` hold type had started to put
a call on MoH hold.

This was done so, to fulfill momentary demands
of call scenario which is in fact wrong and
should not be supported:
- A puts on sendonly hold B
- B puts vice-versa A on inactive hold
- now two-way MoH stream both sides
- B resumes his stream with recvonly (meanwhile
  only sendrecv considered as the only possible
  type to put off hold)

This commits keeps support of stream types
detection as it was, hence `RecvonlyStream` is
still supported. But not anymore considered as
a type which can trigger MoH, in other words
it doesn't put on hold.

This approach makes SEMS in sync with MoH
implementation in rtpengine, whereas only
sendonly and inactive states are allowed to trigger
MoH.

Change-Id: Icf5aa082092cd8f79e43d30339ae8f98a25a7381