MT#62063 rework `CallLeg::updateHoldMethod()`
Rework this function so, that it has all
required for it handling in itself, and doesn't
depend on pre-calls of other things before.

Defined exact purpose of it:
- return `holdMethod` based on SDP
- update `hold_type_requested` class member

It uses `isHoldRequest()` to set the `holdMethod`
but doesn't take into account true/false returned
from it, hence uses this function only as a setter.

Change-Id: Ia89eb20294a2a2533b5fb1e6f5d6d95cbb6179dc