MT#61878 trixie: eliminate implicit operations: enum vs float
C++20 deprecates implicit conversion of enumerations
during math operations like division.

Instead deprecate enum CPS_SAMPLERATE which is used
in a role of pre-proc definition and add a `const float`
value of which is defined by the constructor and is
unchangable.

Also don't use straight-forward (old style) cast with `(<type>)`
but use C++ designed type casts like a static cast, for the
rest of operands in the formula.

Change-Id: Idbfff86f4ec805f3f6db144f3aa041f8fe2c017a