MT#61912 AmB2BSession: move `established_body` to dlg level
Move it to the dlg level, because of complex visibility.
In other words, when working on the dialog level and e.g.
generating new SDP body using `dlg.oa`, there is no way
to save it.

Moving it to the dlg level, gives a control over it on
the whole amount of inheritance levels plus dlg.

Change-Id: Iebb13e3e58f2558dc060e23270d99d690fa54572