MT#61912 OA: save the established body when it's generated
When OA generates own SDP body (e.g. case with
accepted invite in DSP), save it as established_body
for this leg, in order to be able to re-invite
this leg later.

It can happen there is still no SDP seen from the
callee side (other side), and in case we want to
send established re-invite towards this leg,
this will fail, since no SDP body seen yet.

Change-Id: Ifd9f0fb70d27deac871de4eed1648f7c152813f3