MT#62181 AmThread: Initialize `t` static member outside of class
For the `AmThreadLocalStorage` its static `t` member
must be explicitely initialized outside of the template class.

Because it is a template class, it will only be instantiated
when used (when the data type is known), if no initialization
of the static member provided, then linker will fail to link
it (for other users like `RedisCheckpointWriter` which instantiates
`AmThreadLocalStorage` with `RedisConnection`).

Fixes:

ERROR: opening plugin '/usr/lib/sems-pbx/plug-in//redis_store.so':
        /usr/lib/sems-pbx/plug-in//redis_store.so: undefined symbol: _ZN20AmThreadLocalStorageI15RedisConnectionE1tE
ERROR: while loading plug-in '/usr/lib/sems-pbx/plug-in//redis_store.so'
Change-Id: Iad3072c07fd04c5341838bb2a2c08858ca94b03a