MT#62181 AmThread: streamline
Use an atomic enum to track state: idle, running, stopping, or stopped.

Use atomic compare and exchange to ensure consistency without requiring
a mutex.

Never detach threads. Always require a join. This makes for an easy and
reliable way to wait for a thread to finish.

Provide a default on_stop so that it can be omitted in child classes.

Change-Id: Id59942638b7f769090ac91c2049b7d934e876ffc