MT#62181 AmThreadWatcher: overhaul
Use a std::list instead of a std::queue so that we can use iterators,
which allows us to delete elements directly without having to rebuild a
new queue with every loop iteration. This eliminates the n_thread_queue
and the swap.

Eliminate the wait condition _run_cond. This is only woken up when a
thread is added, which we don't care about. We only care when a thread
finishes, and there is no condition for that.

Use a lock guard. With this we can make sure the lock is in a consistent
state in case of an exception. Still catch and report exceptions.

Remove stub methods.

Change-Id: I59903f6a03863e5c106631e3db8706109915c8fb