MT#62181 RegisterCache: make reg_expire private
Provide accessor methods. Put the setter into AorHash and require
iterators. All bindings must be in an AorEntry container and all
AorEntry objects must be in the AorHash container. By putting the setter
in AorHash and using iterators we can make sure that the binding is in
the container already and that expiry values are kept consistent.

Change-Id: I1c5140d1a7547059f61e47cd317680fcb6eae9fe