MT#62181 RegisterCache: add bindings set
Make AorEntry keep an ordered set of all bindings, ordered by expire
time. This makes it easy to quickly tell when the next binding is set to
expire, and which one. Store iterators to bindings in the set so that
they can be used to directly remove the binding when expired.

Change-Id: Ib63c15d5f6e8990e2f4d237726be54da76194a20