MT#62181 DBRegAgent: extend lock span
Temporarily releasing the lock only to immediately re-acquire it, just
because the called function does a lock/unlock, without re-checking all
data that is protected by the lock, opens the door to subtle race
conditions. Change the semantics of removeRegistration() to require
being called with the lock held. Update handleRegistrationTimer() to
also not release and re-acquire the lock.

Change-Id: Ib7562ae138b55d0ef83ec142706ec6415037fc2a