MT#59962 tcp_trsp_socket: `on_read()` protect against data overflow
Refactor it so, that it's not possible to go over what
`unsigned long` offers (so 4,294,967,295).

We are accumulating read bytes there, which can potentially
be more than proposed 4,3Gb, for that case check out if
the possible end result will overflow and close connection
if this is likely to happen.

Change-Id: I4e758292e77649ddfbcaaa7f8dc9f133e74c4600