MT#63051 debian: Fix make variable usage
These variables used to be shell variables passed to the make
invocations in various rules. But those got refactored into make
variables. The problem is that in make the equal sign does not bind
according to surrounding lack of spaces, and the first equal sign
will mark an assignment for the rest of the string including subsequent
equal signs.

The other problem is with the CPPFLAGS assignment where it will assign
append the value and then subsequent conditional assignments will not
take place. Switch to use DEB_CPPFLAGS_APPEND, which was designed
precisely for this purpose.

Fixes: commit 9b5df2495e1fb8ad332895a1cf4e041112da51c6
Change-Id: Ice82295903729fed3f0fb11f750f3f2fdde2e09f