MT#62859 CallLeg: set sip replay to false (reconnect)
When this leg being reconnected, is still in progress
of posting remaining events to update the other side
(the reconnect originator), don't consider it as
sip relay ready. Otherwise any INVITE updates coming
to it, will lead to stumbled INVITEs coming to the other leg,
but without a possibility to use it here (because e.g. this
other side never seen 200OK answer even).

Change-Id: I183d0f70433c156030de38238db395e2c1372352