MT#62181 rework singleton
Reverse inheritence between singleton class and its base.

Change singleton template so that the singleton class doesn't inherit
from its base, but rather that the singelton class inherits from the
singleton template. This removes the need to keep the singleton base
classes separate (with the underscore prefix) plus a `typedef` for the
singleton class, and instead makes the class itself become the
singleton. (The exception being intermediate classes that have multiple
other derived classes, which is only the wheeltimer).

This makes for cleaner typing, but requires use of listing the singleton
template as friend class.

No functional changes.

Change-Id: Ic45cb01f7870ce0ba97188e58340b10fdc0380cb