MT#62859 AmB2BSession::onB2BEvent send fake SDP answer
In case of INVITE pending UAC transactions for the other leg,
and new coming in (like 200OK send to this leg, and just right
after that re-INVITE), just provide a fake SDP answer (200OK)
in the meanwhile.

And expect that a finished UAC transaction on the other side,
must trigger a consequent update for this leg, what will finish
these media negotiations.

Change-Id: I190dfcd5467c728e2192940dfac8454f708cecaf