MT#59962 Rework AmSdp parsing and func users accordingly
Preamble: no functional changes.

AmSdp::parse() mixes int values with bool,
what makes the behavior not really clear, not defined
and can lead to the unpredicted result (even though the
compiler *should* actually fix that, but one ought to
not rely on this).

Hence make the parse method only be working with bool values.

Then, reverse the return value so that it's clear
that wrongly parsed is `false`, and good parsed is `true`.
Refactor all users accordingly.

Also rework those users, who indeed need `int` value
be returned, e.g. when having -1.

Do the same thing with the `parse_sdp_line_ex()` helper,
whereas parse() is the only user of it. And also reverse
the interpretation of true/false.

Then, make `parse()` working with a plain `char*`, so not a const.
Because doing a tricks like:
cast `char*` to const (user level) -> cast `const char*` to non-const
in a function (in a old C-like manner), has really no sense.
Refactor everything accordingly.

Other than that, refactor AmMimeBody:

Refactor it to work with a plain `char *` pointer instead of working
with `unsigned char *`, which nowadays has really no sense and
rather is a rudiment of C-like code base coming from the past.

Convert `payload` from `unsigned char*` to `char *` accordingly.

Refactor everything in AmMimeBody implementation accordingly.

Remove rudiment C-like casting everywhere, where possible.

Update `parseMultipart()` to work with a plain `const char *`
instead of `const unsigned char*`

P.S.: leave a list of TODO's for further rework, which
is not directly related to this scope of rework.

Change-Id: Ie1e132429245e0d2cc740d5b1c1fc17cf037a820