MT#62181 wheeltimer: prevent double free
Requesting removal of a timer that isn't present in the list indicates a
race condition. Something else has already removed/disarmed the timer
(possibly from firing the timer), and that something else is then in
charge of freeing the object. Therefore do not free the timer object at
removal if the timer wasn't armed.

Change-Id: I26b58b6dc3400acf3241375da3cde879a5968d41